+++ src/repositories/controllers.ts
@@ -1,3 +1,4 @@
+import { inAppDirs } from "@src/support/fileWatcher";
 import * as fs from "fs";
 import { repository } from ".";
 import { projectPath } from "../support/project";
@@ -74,12 +75,12 @@ const collectControllers = (path: string): string[] => {
     return [...controllers];
 };
 
-export const getControllers = repository<string[]>(
-    () => {
+export const getControllers = repository<string[]>({
+    load: () => {
         return new Promise((resolve) => {
             resolve(load());
         });
     },
-    "{,**/}{Controllers}{.php,/*.php,/**/*.php}",
-    [],
-);
+    pattern: inAppDirs("{,**/}{Controllers}{.php,/*.php,/**/*.php}"),
+    itemsDefault: [],
+});